factored out InOtherThread to be usable by other tests.
